benchmark/radosbench.py: return 255 if no_version
there is chance that developer compiles ceph with
ENABLE_GIT_VERSION=OFF, in that case `no_version` is printed in the
place of version number in the output of `rados -v`. so "255" is
returned in that case. as it should be large enough to cover the
features we want to use in radosbench.

Signed-off-by: Kefu Chai <kchai@redhat.com>